Vec3 fixes
